refactor: organize sports components into subdirectories
Group 35 flat files in components/sports/ by concern:
- admin/ — admin UI, sidebar, access requests, settings guard, admin-tabs/
- session/ — session cards, dialogs, forms, filters, views, timers
- signup/ — signup buttons, banners, signups table
- root — shared layout components (auth, badges, headers, page shells)

All imports updated to new paths. Zero type errors.
