+++ app/[sport]/page.tsx
@@ -2,13 +2,13 @@ import { Suspense } from "react";
 import { notFound } from "next/navigation";
 import { createClient } from "@/lib/supabase/server";
 import { getUser, getUserSportRole } from "@/lib/supabase/user";
-import SessionCard from "@/components/sports/session-card";
-import SessionFilter from "@/components/sports/session-filter";
-import TeamAccessBanner from "@/components/sports/team-access-banner";
-import SignInToSignupBanner from "@/components/sports/sign-in-to-signup-banner";
+import SessionCard from "@/components/sports/session/session-card";
+import SessionFilter from "@/components/sports/session/session-filter";
+import TeamAccessBanner from "@/components/sports/signup/team-access-banner";
+import SignInToSignupBanner from "@/components/sports/signup/sign-in-to-signup-banner";
 import SportPageShell from "@/components/sports/sport-page-shell";
-import AdminButton from "@/components/sports/admin-button";
-import CalendarExportButton from "@/components/sports/calendar-export-button";
+import AdminButton from "@/components/sports/admin/admin-button";
+import CalendarExportButton from "@/components/sports/session/calendar-export-button";
 import { Role, AccessLevel, getResolvedTab } from "@/config/config-resolver";
 import type { SessionTab, AccessBannerText, ResolvedSportConfig } from "@/config/config-resolver";
 import { getFirstUnmetLevel, ACCESS_LEVELS } from "@/lib/tab-access";
