+++ components/sports/admin/admin-tabs/settings/admin-tab-ui-metadata.ts
