+++ components/sports/session/delete-session-button.tsx
